build: self-contained makefile, direct baserom incbin, and setup.sh
- Makefile no longer depends on ../pokeemerald: uses tools/agbcc (built by
  setup.sh from pret/agbcc cloned to tools/agbcc-src) for CC and headers
- C_SRCS now uses wildcard src/*.c src/*/*.c with UNPORTED_SRCS exclude
  list (28 US reference files still living in asm); consolidated the ~20
  duplicate per-file compile rules into one generic rule with per-target
  CFLAGS overrides
- data/*.s, data_rest.s, event_scripts.s incbin directly from
  baserom_jp.gba at absolute ROM offsets; removed
  tools/extract_baserom_data.py and the build/data/*.bin extraction step
- multiboot programs extracted to data/mb_ereader.gba and
  data/mb_berry_fix.gba (committed, like pokeemerald); .s files incbin
  them directly; .gitignore whitelists data/mb_*.gba and ignores
  tools/agbcc-src + baserom_jp.gba
- setup.sh: one-shot env setup (system tool check, baserom SHA1 check,
  build_tools.sh, binutils links, agbcc clone/build/install); README updated
- Fix type conflicts the generic compile rule exposed: money.h JP
  signatures (PrintMoneyAmount u16 x, no speed; MoneyBoxWithBorder speed),
  gPokenavOptions_Tilemap u16, MatchCall/Jp prototypes, palette.c orphaned
  array initializer, script_movement gBitTable, etc.
- SHA1 verified byte-exact (sha1sum -c rom_jp.sha1)

decomp progress: 3435/12612 = 27.24% (asm -> C)

+++ src/pokenav_conditions.c
@@ -63,7 +63,7 @@ extern const struct BgTemplate sMenuBgTemplates[3];
 extern const u32 gPokenavCondition_Gfx[];
 extern const u32 sConditionGraphData_Gfx[];
 extern const u32 gPokenavCondition_Tilemap[];
-extern const u32 gPokenavOptions_Tilemap[];
+extern const u16 gPokenavOptions_Tilemap[];
 extern const u16 gPokenavCondition_Pal[];
 extern const u16 gConditionText_Pal[];
 extern const u32 sConditionGraphData_Tilemap[];
